♻️ refactor(ui): update on get universities call
Ensures get university is only called once on the base onboardingscreen widget. This  involves making use of a notifier and a listener (watcher) in updating the university list wherever it is used. The change also implements a single instance of the Global Storage class by providing it inside the build class wherever it is needed.